test-requirements.txt
nose2
pyyaml
webtest
